backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Navarro-Jonathan <navarrojonathanj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^8.2.0"
Flask = "^3.0.0"
Flask-Cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
